refactor(rtk-query): rename emptyProfile file and fix SVG strokeWidth attribute
